+++ actix-web-lab/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased - 2022-xx-xx
+- Implement `FromStr` and `Header::parse` for `StrictTransportSecurity`.
+- Implement `PartialEq` and `Eq` for `StrictTransportSecurity`.
 
 
 ## 0.16.6 - 2022-07-27